pyproject.toml
[project]
name = "decp.info"
description = "Interface d'exploration et d'analyse des marchés publics français."
version = "2.7.4"
requires-python = ">= 3.10"
authors = [{ name = "Colin Maudry", email = "colin@colmo.tech" }]
dependencies = [
  "dash==3.4.0",
  "dash[compress]",
  "polars",
  "gunicorn",
  "dash-bootstrap-components",
  "python-dotenv",
  "xlsxwriter",
  "plotly[express]",
  "httpx",
  "pandas",                    # utilisé pour la création de certains graphiques
  "unidecode",
  "dash-leaflet",
  "dash-extensions",
  "duckdb",
  "flask-caching",
  "pyarrow>=23.0.1",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-env",
  "pre-commit",
  "selenium",
  "webdriver-manager",
  "dash[testing]",
  "fastexcel",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
env = [
  "DATA_FILE_PARQUET_PATH=tests/test.parquet",
  "DEVELOPMENT=true",
  "REBUILD_DUCKDB=true",
  "DATA_SCHEMA_PATH=/home/colin/git/decp-processing/dist/schema.json",
]
addopts = "-p no:warnings"
